Add delete confirm

+++ public/ng/controllers/humans/listHumanController.js
@@ -6,13 +6,17 @@ function listHumanController ($scope, $apiRequest, $init, $page) {
     $scope.humans = $init;
 
     //Delete Human
-    $scope.deleteHumanController= function (id) {
-        $apiRequest.config( {
-            method: "DELETE",
-            url: "humans/" + id,
-        }, function (response, data) {
-            $page.reload();
-        }).send();
+    $scope.deleteHumanController = function (id) {
+        var result = confirm("هل تريد الحذف؟");
+        if (result) {
+            $apiRequest.config( {
+                method: "DELETE",
+                url: "humans/" + id,
+            }, function (response, data) {
+                $page.reload();
+            }).send();
+        }
+
     }
 }
